+++ src/Akeneo/Pim/Enrichment/Bundle/Elasticsearch/Filter/Attribute/ReferenceDataFilter.php
@@ -7,7 +7,7 @@ use Akeneo\Pim\Enrichment\Component\Product\Exception\InvalidOperatorException;
 use Akeneo\Pim\Enrichment\Component\Product\Query\Filter\AttributeFilterInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Query\Filter\FieldFilterHelper;
 use Akeneo\Pim\Enrichment\Component\Product\Query\Filter\Operators;
-use Akeneo\Pim\Enrichment\Component\Product\Validator\AttributeValidatorHelper;
+use Akeneo\Pim\Enrichment\Component\Product\Validator\ElasticsearchFilterValidator;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Akeneo\Pim\Structure\Component\ReferenceData\ConfigurationRegistryInterface;
 use Akeneo\Tool\Component\StorageUtils\Exception\InvalidPropertyException;
@@ -27,21 +27,14 @@ class ReferenceDataFilter extends AbstractAttributeFilter implements AttributeFi
     /** @var ConfigurationRegistryInterface */
     protected $registry;
 
-    /**
-     * @param AttributeValidatorHelper         $attrValidatorHelper
-     * @param ReferenceDataRepositoryResolver  $referenceDataRepositoryResolver
-     * @param ConfigurationRegistryInterface   $registry
-     * @param array                            $supportedAttributeTypes
-     * @param array                            $supportedOperators
-     */
     public function __construct(
-        AttributeValidatorHelper $attrValidatorHelper,
+        ElasticsearchFilterValidator $filterValidator,
         ReferenceDataRepositoryResolver $referenceDataRepositoryResolver,
         ConfigurationRegistryInterface $registry,
         array $supportedAttributeTypes = [],
         array $supportedOperators = []
     ) {
-        $this->attrValidatorHelper = $attrValidatorHelper;
+        $this->filterValidator = $filterValidator;
         $this->referenceDataRepositoryResolver = $referenceDataRepositoryResolver;
         $this->supportedAttributeTypes = $supportedAttributeTypes;
         $this->supportedOperators = $supportedOperators;