+++ tests/back/Pim/Enrichment/Specification/Bundle/Elasticsearch/Filter/Attribute/MediaFilterSpec.php
@@ -11,14 +11,14 @@ use Akeneo\Pim\Enrichment\Component\Product\Exception\InvalidOperatorException;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Query\Filter\AttributeFilterInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Query\Filter\Operators;
-use Akeneo\Pim\Enrichment\Component\Product\Validator\AttributeValidatorHelper;
+use Akeneo\Pim\Enrichment\Component\Product\Validator\ElasticsearchFilterValidator;
 
 class MediaFilterSpec extends ObjectBehavior
 {
-    function let(AttributeValidatorHelper $attributeValidatorHelper)
+    function let(ElasticsearchFilterValidator $filterValidator)
     {
         $this->beConstructedWith(
-            $attributeValidatorHelper,
+            $filterValidator,
             ['pim_catalog_file', 'pim_catalog_image'],
             ['STARTS WITH', 'CONTAINS', 'DOES NOT CONTAIN', '=', '!=', 'EMPTY', 'NOT EMPTY']
         );
@@ -50,15 +50,15 @@ class MediaFilterSpec extends ObjectBehavior
     }
 
     function it_adds_a_filter_with_operator_starts_with(
-        $attributeValidatorHelper,
+        $filterValidator,
         AttributeInterface $name,
         SearchQueryBuilder $sqb
     ) {
         $name->getCode()->willReturn('an_image');
         $name->getBackendType()->willReturn('media');
 
-        $attributeValidatorHelper->validateLocale($name, 'en_US')->shouldBeCalled();
-        $attributeValidatorHelper->validateScope($name, 'ecommerce')->shouldBeCalled();
+        $filterValidator->validateLocaleForAttribute('an_image', 'en_US')->shouldBeCalled();
+        $filterValidator->validateChannelForAttribute('an_image', 'ecommerce')->shouldBeCalled();
 
         $sqb->addFilter(
             [
@@ -74,15 +74,15 @@ class MediaFilterSpec extends ObjectBehavior
     }
 
     function it_adds_a_filter_with_operator_contains(
-        $attributeValidatorHelper,
+        $filterValidator,
         AttributeInterface $name,
         SearchQueryBuilder $sqb
     ) {
         $name->getCode()->willReturn('an_image');
         $name->getBackendType()->willReturn('media');
 
-        $attributeValidatorHelper->validateLocale($name, 'en_US')->shouldBeCalled();
-        $attributeValidatorHelper->validateScope($name, 'ecommerce')->shouldBeCalled();
+        $filterValidator->validateLocaleForAttribute('an_image', 'en_US')->shouldBeCalled();
+        $filterValidator->validateChannelForAttribute('an_image', 'ecommerce')->shouldBeCalled();
 
         $sqb->addFilter(
             [
@@ -98,15 +98,15 @@ class MediaFilterSpec extends ObjectBehavior
     }
 
     function it_adds_a_filter_with_operator_does_not_contain(
-        $attributeValidatorHelper,
+        $filterValidator,
         AttributeInterface $name,
         SearchQueryBuilder $sqb
     ) {
         $name->getCode()->willReturn('an_image');
         $name->getBackendType()->willReturn('media');
 
-        $attributeValidatorHelper->validateLocale($name, 'en_US')->shouldBeCalled();
-        $attributeValidatorHelper->validateScope($name, 'ecommerce')->shouldBeCalled();
+        $filterValidator->validateLocaleForAttribute('an_image', 'en_US')->shouldBeCalled();
+        $filterValidator->validateChannelForAttribute('an_image', 'ecommerce')->shouldBeCalled();
 
         $sqb->addMustNot(
             [
@@ -130,15 +130,15 @@ class MediaFilterSpec extends ObjectBehavior
     }
 
     function it_adds_a_filter_with_operator_equals(
-        $attributeValidatorHelper,
+        $filterValidator,
         AttributeInterface $name,
         SearchQueryBuilder $sqb
     ) {
         $name->getCode()->willReturn('an_image');
         $name->getBackendType()->willReturn('media');
 
-        $attributeValidatorHelper->validateLocale($name, 'en_US')->shouldBeCalled();
-        $attributeValidatorHelper->validateScope($name, 'ecommerce')->shouldBeCalled();
+        $filterValidator->validateLocaleForAttribute('an_image', 'en_US')->shouldBeCalled();
+        $filterValidator->validateChannelForAttribute('an_image', 'ecommerce')->shouldBeCalled();
 
         $sqb->addFilter(
             [
@@ -153,15 +153,15 @@ class MediaFilterSpec extends ObjectBehavior
     }
 
     function it_adds_a_filter_with_operator_not_equal(
-        $attributeValidatorHelper,
+        $filterValidator,
         AttributeInterface $name,
         SearchQueryBuilder $sqb
     ) {
         $name->getCode()->willReturn('an_image');
         $name->getBackendType()->willReturn('media');
 
-        $attributeValidatorHelper->validateLocale($name, 'en_US')->shouldBeCalled();
-        $attributeValidatorHelper->validateScope($name, 'ecommerce')->shouldBeCalled();
+        $filterValidator->validateLocaleForAttribute('an_image', 'en_US')->shouldBeCalled();
+        $filterValidator->validateChannelForAttribute('an_image', 'ecommerce')->shouldBeCalled();
 
         $sqb->addMustNot(
             [
@@ -182,15 +182,15 @@ class MediaFilterSpec extends ObjectBehavior
     }
 
     function it_adds_a_filter_with_operator_empty(
-        $attributeValidatorHelper,
+        $filterValidator,
         AttributeInterface $name,
         SearchQueryBuilder $sqb
     ) {
         $name->getCode()->willReturn('an_image');
         $name->getBackendType()->willReturn('media');
 
-        $attributeValidatorHelper->validateLocale($name, 'en_US')->shouldBeCalled();
-        $attributeValidatorHelper->validateScope($name, 'ecommerce')->shouldBeCalled();
+        $filterValidator->validateLocaleForAttribute('an_image', 'en_US')->shouldBeCalled();
+        $filterValidator->validateChannelForAttribute('an_image', 'ecommerce')->shouldBeCalled();
 
         $sqb->addMustNot(
             [
@@ -206,15 +206,15 @@ class MediaFilterSpec extends ObjectBehavior
     }
 
     function it_adds_a_filter_with_operator_not_empty(
-        $attributeValidatorHelper,
+        $filterValidator,
         AttributeInterface $name,
         SearchQueryBuilder $sqb
     ) {
         $name->getCode()->willReturn('an_image');
         $name->getBackendType()->willReturn('media');
 
-        $attributeValidatorHelper->validateLocale($name, 'en_US')->shouldBeCalled();
-        $attributeValidatorHelper->validateScope($name, 'ecommerce')->shouldBeCalled();
+        $filterValidator->validateLocaleForAttribute('an_image', 'en_US')->shouldBeCalled();
+        $filterValidator->validateChannelForAttribute('an_image', 'ecommerce')->shouldBeCalled();
 
         $sqb->addFilter(
             [
@@ -236,15 +236,15 @@ class MediaFilterSpec extends ObjectBehavior
     }
 
     function it_throws_an_exception_when_the_given_value_is_not_a_string(
-        $attributeValidatorHelper,
+        $filterValidator,
         AttributeInterface $name,
         SearchQueryBuilder $sqb
     ) {
         $name->getCode()->willReturn('an_image');
         $name->getBackendType()->willReturn('media');
 
-        $attributeValidatorHelper->validateLocale($name, 'en_US')->shouldBeCalled();
-        $attributeValidatorHelper->validateScope($name, 'ecommerce')->shouldBeCalled();
+        $filterValidator->validateLocaleForAttribute('an_image', 'en_US')->shouldBeCalled();
+        $filterValidator->validateChannelForAttribute('an_image', 'ecommerce')->shouldBeCalled();
 
         $this->setQueryBuilder($sqb);
 
@@ -258,15 +258,15 @@ class MediaFilterSpec extends ObjectBehavior
     }
 
     function it_throws_an_exception_when_it_filters_on_an_unsupported_operator(
-        $attributeValidatorHelper,
+        $filterValidator,
         AttributeInterface $name,
         SearchQueryBuilder $sqb
     ) {
         $name->getCode()->willReturn('an_image');
         $name->getBackendType()->willReturn('media');
 
-        $attributeValidatorHelper->validateLocale($name, 'en_US')->shouldBeCalled();
-        $attributeValidatorHelper->validateScope($name, 'ecommerce')->shouldBeCalled();
+        $filterValidator->validateLocaleForAttribute('an_image', 'en_US')->shouldBeCalled();
+        $filterValidator->validateChannelForAttribute('an_image', 'ecommerce')->shouldBeCalled();
 
         $this->setQueryBuilder($sqb);
 
@@ -279,7 +279,7 @@ class MediaFilterSpec extends ObjectBehavior
     }
 
     function it_throws_an_exception_when_an_exception_is_thrown_by_the_attribute_validator_on_locale_validation(
-        $attributeValidatorHelper,
+        $filterValidator,
         AttributeInterface $name,
         SearchQueryBuilder $sqb
     ) {
@@ -289,7 +289,7 @@ class MediaFilterSpec extends ObjectBehavior
         $name->getAvailableLocaleCodes('fr_FR');
 
         $e = new \LogicException('Attribute "name" expects a locale, none given.');
-        $attributeValidatorHelper->validateLocale($name, 'en_US')->willThrow($e);
+        $filterValidator->validateLocaleForAttribute('an_image', 'en_US')->willThrow($e);
 
         $this->setQueryBuilder($sqb);
 
@@ -303,7 +303,7 @@ class MediaFilterSpec extends ObjectBehavior
     }
 
     function it_throws_an_exception_when_an_exception_is_thrown_by_the_attribute_validator_on_scope_validation(
-        $attributeValidatorHelper,
+        $filterValidator,
         AttributeInterface $name,
         SearchQueryBuilder $sqb
     ) {
@@ -312,7 +312,7 @@ class MediaFilterSpec extends ObjectBehavior
         $name->isScopable()->willReturn(false);
 
         $e = new \LogicException('Attribute "name" does not expect a scope, "ecommerce" given.');
-        $attributeValidatorHelper->validateLocale($name, 'en_US')->willThrow($e);
+        $filterValidator->validateLocaleForAttribute('an_image', 'en_US')->willThrow($e);
 
         $this->setQueryBuilder($sqb);