+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - PIM-9119: Fix missing warning when using mass edit with parent filter set to empty
 - PIM-9114: fix errors on mass action when the parent filter is set to empty
 - PIM-9110: avoid deadlock error when laoding products in parallel with the API
+- PIM-9113: Locale Specific attribute breaks product grid